Move `create_expanded_workflow_graph` into tool_manager.utils
